adding support for a record set collection view
This component allows you to visualize a record set object. This provides support for pagination. As the selected page of the record set changes the view is updated. This allows decoupled control over recordset population, pagination, and view resolution. For example you could load pages of commetns into a record set object and bing the recordset collection view instance to the data in the model. As the user pages through the results in the model object the view will update.